refactor(dap): split DAP protocol types into microcrate
