+++ Cargo.toml
@@ -75,6 +75,7 @@ members = [
     "crates/perl-corpus",
     "crates/perl-lsp",
     "crates/perl-dap",
+    "crates/perl-dap-protocol",
     "crates/perl-dead-code",
     "crates/perl-dap-breakpoint",
     "crates/perl-dap-eval",
@@ -163,6 +164,7 @@ allow = [
 
     # Tier 5 — DAP components
     "perl-dap-breakpoint",
+    "perl-dap-protocol",
     "perl-dap-eval",
     "perl-dap-platform",
     "perl-dap-stack",
@@ -260,6 +262,7 @@ perl-corpus = { path = "crates/perl-corpus", version = "0.10.0" }
 perl-dap = { path = "crates/perl-dap", version = "0.10.0" }
 perl-dead-code = { path = "crates/perl-dead-code", version = "0.10.0" }
 perl-dap-breakpoint = { path = "crates/perl-dap-breakpoint", version = "0.10.0" }
+perl-dap-protocol = { path = "crates/perl-dap-protocol", version = "0.10.0" }
 perl-dap-eval = { path = "crates/perl-dap-eval", version = "0.10.0" }
 perl-dap-platform = { path = "crates/perl-dap-platform", version = "0.10.0" }
 perl-dap-variables = { path = "crates/perl-dap-variables", version = "0.10.0" }
