+++ crates/perl-dap-protocol/README.md
@@ -1,0 +1,6 @@
+# perl-dap-protocol
+
+Shared Debug Adapter Protocol (DAP) type definitions for the perl-lsp workspace.
+
+This crate contains serializable request/response/event structures used by
+`perl-dap` and related tooling.
